Issue #3
Create a class called CMV to store the result from each LIC.
Furthermore, add a test file for the CMV class.
Lastly, add all the necessary classes needed by the CMV
class which are Parameters and Point. In the future the CMV
class might need more classes such as a Line class and a Circle class.