Rewrite verification method to be static.
The `Parameters.verify()` needed access to the number of points. This
commit solves that by passing the `points` array as an argument to the
`verify`-method, instead of requiring NUMPOINTS to be a field in the
`Parameters` class.

The test cases have been rewritten to reflect the changes, and have also
gotten better documentation.